Start error handeling

+++ src/com/vojat/Main.java
@@ -54,7 +54,7 @@ public class Main {
             Game.langFileName = jsonEditor.readData("Language");
             Game.texturePack = jsonEditor.readData("Texture-Pack");
             Game.version = jsonEditor.readData("Version");
-            Game.FPS_SET = Byte.parseByte(jsonEditor.readData("FPS"));
+            Game.FPS_SET = (byte) Integer.parseInt(jsonEditor.readData("FPS"));
             Game.setDayLasts(Integer.parseInt(jsonEditor.readData("Day-Lasts")));
             Game.setNightLasts(Integer.parseInt(jsonEditor.readData("Night-Lasts")));
             Game.dayNightTransitionSpeed = Double.parseDouble(jsonEditor.readData("Cycle-Transition-Value"));
